order route jwt based authorization added

+++ Controllers/Order.Controller.js
@@ -2,22 +2,36 @@ const createError = require('http-errors');
 const mongoose = require('mongoose');
 
 const Order = require('../Models/Order.model');
+const verifyToken = require('../utils/verifyToken');
+
 
 module.exports = {
     getAllOrders: async (req, res, next) => {
         try {
-            const results = await Order.find({});
-            res.send(results);
+            if (req.decodedEmail) {
+                const results = await Order.find({});
+                res.send(results);
+            } else {
+                res.status(401).json({ message: "User not Authorized" });
+            }
+
         } catch (error) {
             console.log(error.message);
         }
     },
     getAllUserOrders: async (req, res, next) => {
+        const email = req.params.email;
         try {
-            const filter = { 'user.email': req.body.email };
-            console.log(filter);
-            const results = await Order.find(filter);
-            res.send(results);
+            if (email == req.decodedEmail) {
+                const filter = { 'user.email': email };
+                console.log(filter);
+                console.log(req.decodedEmail);
+                const results = await Order.find(filter);
+                res.send(results);
+            } else {
+                res.status(401).json({ message: "User not Authorized" });
+            }
+
         } catch (error) {
             console.log(error.message);
         }
@@ -25,6 +39,13 @@ module.exports = {
 
     createNewOrder: async (req, res, next) => {
         try {
+            if (req.decodedEmail) {
+
+            } else {
+                res.status(401).json({ message: "User not Authorized" });
+
+            }
+
             const order = new Order(req.body);
             const result = await order.save();
             res.send(result);
@@ -41,12 +62,18 @@ module.exports = {
     findOrderById: async (req, res, next) => {
         const id = req.params.id;
         try {
-            const order = await Order.findById(id);
-            // const order = await order.findOne({ _id: id });
-            if (!order) {
-                throw createError(404, 'order does not exist.');
+            if (req.decodedEmail) {
+                const order = await Order.findById(id);
+                // const order = await order.findOne({ _id: id });
+                if (!order) {
+                    throw createError(404, 'order does not exist.');
+                }
+                res.send(order);
+            } else {
+                res.status(401).json({ message: "User not Authorized" });
+
             }
-            res.send(order);
+
         } catch (error) {
             console.log(error.message);
             if (error instanceof mongoose.CastError) {
@@ -59,15 +86,21 @@ module.exports = {
 
     updateAOrder: async (req, res, next) => {
         try {
-            const id = req.params.id;
-            const updates = req.body;
-            const options = { new: false };
+            if (req.decodedEmail) {
+                const id = req.params.id;
+                const updates = req.body;
+                const options = { new: false };
 
-            const result = await Order.findByIdAndUpdate(id, updates, options);
-            if (!result) {
-                throw createError(404, 'Order does not exist');
+                const result = await Order.findByIdAndUpdate(id, updates, options);
+                if (!result) {
+                    throw createError(404, 'Order does not exist');
+                }
+                res.send(result);
+            } else {
+                res.status(401).json({ message: "User not Authorized" });
+
             }
-            res.send(result);
+
         } catch (error) {
             console.log(error.message);
             if (error instanceof mongoose.CastError) {
@@ -81,11 +114,19 @@ module.exports = {
     deleteAOrder: async (req, res, next) => {
         const id = req.params.id;
         try {
-            const result = await Order.findByIdAndDelete(id);
-            if (!result) {
-                throw createError(404, 'Order does not exist.');
+            if (req.decodedEmail) {
+                const result = await Order.findByIdAndDelete(id);
+                if (!result) {
+                    throw createError(404, 'Order does not exist.');
+                }
+                res.send(result);
+            } else {
+                res.status(401).json({ message: "User not Authorized" });
+
             }
-            res.send(result);
+
+
+
         } catch (error) {
             console.log(error.message);
             if (error instanceof mongoose.CastError) {
